Extract GCSFS to interface (#770)
(cherry picked from commit 111d06c48fdab0280e0aa4a379cfde9e069943c3)
