+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/InMemoryGoogleHadoopFileSystem.java
@@ -17,6 +17,7 @@ package com.google.cloud.hadoop.fs.gcs;
 import static com.google.cloud.hadoop.gcsio.testing.InMemoryGoogleCloudStorage.getInMemoryGoogleCloudStorageOptions;
 
 import com.google.cloud.hadoop.gcsio.GoogleCloudStorageFileSystem;
+import com.google.cloud.hadoop.gcsio.GoogleCloudStorageFileSystemImpl;
 import com.google.cloud.hadoop.gcsio.GoogleCloudStorageFileSystemOptions;
 import com.google.cloud.hadoop.gcsio.testing.InMemoryGoogleCloudStorage;
 import java.io.IOException;
@@ -48,7 +49,7 @@ public class InMemoryGoogleHadoopFileSystem extends GoogleHadoopFileSystem {
       return inMemoryGcsFs;
     }
     try {
-      return new GoogleCloudStorageFileSystem(
+      return new GoogleCloudStorageFileSystemImpl(
           InMemoryGoogleCloudStorage::new,
           GoogleCloudStorageFileSystemOptions.builder()
               .setCloudStorageOptions(getInMemoryGoogleCloudStorageOptions())
