+++ gcsio/src/test/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageFileSystemIntegrationHelper.java
@@ -38,7 +38,7 @@ public class GoogleCloudStorageFileSystemIntegrationHelper
             .setCopyWithRewriteEnabled(true)
             .build();
 
-    return new GoogleCloudStorageFileSystem(
+    return new GoogleCloudStorageFileSystemImpl(
         GoogleCloudStorageTestHelper.getCredential(),
         GoogleCloudStorageFileSystemOptions.builder()
             .setBucketDeleteEnabled(true)
@@ -178,6 +178,6 @@ public class GoogleCloudStorageFileSystemIntegrationHelper
   }
 
   public String getItemName(URI src) {
-    return gcsfs.getItemName(src);
+    return GoogleCloudStorageFileSystemImpl.getItemName(src);
   }
 }
