+++ gcsio/src/test/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageFileSystemIntegrationTest.java
@@ -135,7 +135,7 @@ public class GoogleCloudStorageFileSystemIntegrationTest {
                                 .build())
                         .build());
 
-            gcsfs = new GoogleCloudStorageFileSystem(credential, optionsBuilder.build());
+            gcsfs = new GoogleCloudStorageFileSystemImpl(credential, optionsBuilder.build());
 
             gcs = gcsfs.getGcs();
 
@@ -1677,7 +1677,7 @@ public class GoogleCloudStorageFileSystemIntegrationTest {
   private List<URI> getSubDirPaths(URI path) {
     StorageResourceId resourceId = StorageResourceId.fromUriPath(path, true);
 
-    List<String> subdirs = GoogleCloudStorageFileSystem.getDirs(resourceId.getObjectName());
+    List<String> subdirs = GoogleCloudStorageFileSystemImpl.getDirs(resourceId.getObjectName());
     List<URI> subDirPaths = new ArrayList<>(subdirs.size());
     for (String subdir : subdirs) {
       subDirPaths.add(gcsiHelper.getPath(resourceId.getBucketName(), subdir));
